correct timer

+++ programs/assignment_2/pargrep.c
@@ -266,7 +266,7 @@ int main(int argc, char* argv[]) {
 
     // stop the timer and print runtime
     clock_gettime(CLOCK_MONOTONIC_RAW, &stop);
-    uint64_t delta = (stop.tv_sec - start.tv_sec) * 1000000 + 
+    long long delta = (stop.tv_sec - start.tv_sec) * 1000000 + 
                      (stop.tv_nsec - start.tv_nsec) / 1000;
     printf("Runtime: %ld", uint64_t);
